cast to float added. this avoids problems with very small values

+++ finley/src/finleyC/Mesh_saveVTK.c
@@ -484,24 +484,24 @@ void Finley_Mesh_saveVTK(const char * filename_p, Finley_Mesh *mesh_p, const dim
    	         /* probably only need to get shape of first element */
    	         /* write the data different ways for scalar, vector and tensor */
    	         if (nCompReqd == 1) {
-   	           fprintf(fileHandle_p, " %e", sampleAvg[0]);
+   	           fprintf(fileHandle_p, " %e", (float) sampleAvg[0]);
    	         } else if (nCompReqd == 3) {
    	           /* write out the data */
-   	           for (int m=0; m<shape; m++) fprintf(fileHandle_p, " %e", sampleAvg[m]);
-   	           for (int m=0; m<nCompReqd-shape; m++) fprintf(fileHandle_p, " %e", 0.);
+   	           for (int m=0; m<shape; m++) fprintf(fileHandle_p, " %e", (float) sampleAvg[m]);
+   	           for (int m=0; m<nCompReqd-shape; m++) fprintf(fileHandle_p, " %e", (float) 0.);
    	         } else if (nCompReqd == 9) { 
    	           /* tensor data, so have a 3x3 matrix to output as a row 
    	            * of 9 data points */
    	            int count = 0;
    	            for (int m=0; m<shape; m++) {
    	              for (int n=0; n<shape; n++) {
-   	                 fprintf(fileHandle_p, " %e", sampleAvg[count]);
+   	                 fprintf(fileHandle_p, " %e", (float) sampleAvg[count]);
    	                 count++;
    	              }
-   	              for (int n=0; n<3-shape; n++) fprintf(fileHandle_p, " %e", 0.);
+   	              for (int n=0; n<3-shape; n++) fprintf(fileHandle_p, " %e", (float) 0.);
    	            }
    	            for (int m=0; m<3-shape; m++) 
-   	               for (int n=0; n<3; n++) fprintf(fileHandle_p, " %e", 0.);
+   	               for (int n=0; n<3; n++) fprintf(fileHandle_p, " %e", (float) 0.);
    	            }
    	          fprintf(fileHandle_p, "\n");
              }
@@ -614,23 +614,23 @@ void Finley_Mesh_saveVTK(const char * filename_p, Finley_Mesh *mesh_p, const dim
    	       /* write the data different ways for scalar, vector and tensor */
                if (do_write) {
    	          if (nCompReqd == 1) {
-   	            fprintf(fileHandle_p, " %e", values[0]);
+   	            fprintf(fileHandle_p, " %e", (float) values[0]);
    	          } else if (nCompReqd == 3) {
-   	            for (int m=0; m<shape; m++) fprintf(fileHandle_p, " %e", values[m]);
-   	            for (int m=0; m<nCompReqd-shape; m++) fprintf(fileHandle_p, " %e", 0.);
+   	            for (int m=0; m<shape; m++) fprintf(fileHandle_p, " %e", (float) values[m]);
+   	            for (int m=0; m<nCompReqd-shape; m++) fprintf(fileHandle_p, " %e", (float) 0.);
    	          } else if (nCompReqd == 9) { 
    	            /* tensor data, so have a 3x3 matrix to output as a row 
    	             * of 9 data points */
    	            int count = 0;
    	            for (int m=0; m<shape; m++) {
    	              for (int n=0; n<shape; n++) {
-   	                fprintf(fileHandle_p, " %e", values[count]);
+   	                fprintf(fileHandle_p, " %e", (float) values[count]);
    	                count++;
    	              }
-   	              for (int n=0; n<3-shape; n++) fprintf(fileHandle_p, " %e", 0.);
+   	              for (int n=0; n<3-shape; n++) fprintf(fileHandle_p, " %e", (float) 0.);
    	            }
    	            for (int m=0; m<3-shape; m++)  
-   	                for (int n=0; n<3; n++) fprintf(fileHandle_p, " %e", 0.);
+   	                for (int n=0; n<3; n++) fprintf(fileHandle_p, " %e", (float) 0.);
    	          }
       	          fprintf(fileHandle_p, "\n");
                }
